requirements.txt
pymongo[tls,srv]==3.11.0
numpy==1.19.1
Flask==1.1.2
python-dotenv==0.14.0
